refactor: add "practice-history-chart.tsx"
